+++ src/orders/schemas.py
@@ -1,9 +1,8 @@
 from datetime import datetime
-from decimal import Decimal
 
 from pydantic import BaseModel, ConfigDict, field_validator
 
-from src.orders.models import OrderStatusEnum
+from orders.models import OrderStatusEnum
 
 
 class OrderItemResponse(BaseModel):
@@ -11,12 +10,12 @@ class OrderItemResponse(BaseModel):
 
     movie_id: int
     name: str
-    price_at_order: Decimal
+    price_at_order: float
 
     @field_validator("price_at_order", mode="before")
     @classmethod
-    def _coerce_price(cls, value: Decimal) -> Decimal | None:
-        return Decimal(str(value)) if value is not None else value
+    def _coerce_price(cls, value: float) -> float | None:
+        return float(str(value)) if value is not None else value
 
 
 class OrderResponse(BaseModel):
@@ -24,13 +23,13 @@ class OrderResponse(BaseModel):
     user_id: int
     created_at: datetime
     status: OrderStatusEnum
-    total_amount: Decimal | None
+    total_amount: float | None
     items: list[OrderItemResponse]
 
     @field_validator("total_amount", mode="before")
     @classmethod
-    def _coerce_total_amount(cls, value: Decimal) -> Decimal | None:
-        return Decimal(str(value)) if value is not None else value
+    def _coerce_total_amount(cls, value: float) -> float | None:
+        return float(str(value)) if value is not None else value
 
 
 class ExcludeMovieResponse(BaseModel):
@@ -48,10 +47,10 @@ class OrderListItemResponse(BaseModel):
     id: int
     created_at: datetime
     status: OrderStatusEnum
-    total_amount: Decimal | None
+    total_amount: float | None
     movie_count: int
 
     @field_validator("total_amount", mode="before")
     @classmethod
-    def _coerce_total_amount(cls, value: Decimal) -> Decimal | None:
-        return Decimal(str(value)) if value is not None else value
+    def _coerce_total_amount(cls, value: float) -> float | None:
+        return float(str(value)) if value is not None else value
